+++ internal/forge/gh/connector.go
@@ -1,14 +1,23 @@
 package gh
 
 import (
+	"errors"
+	"fmt"
+	"regexp"
+	"slices"
+	"strings"
+
 	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
 	"github.com/git-town/git-town/v21/internal/forge/github"
+	"github.com/git-town/git-town/v21/internal/messages"
 	"github.com/git-town/git-town/v21/internal/subshell/subshelldomain"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
 )
 
 var (
 	ghConnector Connector
-	_           forgedomain.Connector = ghConnector
+	_           forgedomain.Connector    = ghConnector
+	_           forgedomain.AuthVerifier = ghConnector
 )
 
 // Connector provides standardized connectivity for the given repository (github.com/owner/repo)
@@ -38,4 +47,48 @@ func (self Connector) DefaultProposalMessage(data forgedomain.ProposalData) stri
 
 func (self Connector) OpenRepository(runner subshelldomain.Runner) error {
 	return runner.Run("gh", "browse")
+}
+
+func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
+	output, err := self.Backend.Query("gh", "auth", "status", "--active")
+	if err != nil {
+		return forgedomain.VerifyConnectionResult{
+			AuthenticatedUser:   None[string](),
+			AuthenticationError: err,
+			AuthorizationError:  nil,
+		}
+	}
+	return ParsePermissionsOutput(output)
+}
+
+func ParsePermissionsOutput(output string) forgedomain.VerifyConnectionResult {
+	result := forgedomain.VerifyConnectionResult{
+		AuthenticatedUser:   None[string](),
+		AuthenticationError: nil,
+		AuthorizationError:  nil,
+	}
+	lines := strings.Split(output, "\n")
+	regex := regexp.MustCompile(`Logged in to github.com account (\w+)`)
+	for _, line := range lines {
+		matches := regex.FindStringSubmatch(line)
+		if matches != nil {
+			result.AuthenticatedUser = NewOption(matches[1])
+			break
+		}
+	}
+	if result.AuthenticatedUser.IsNone() {
+		result.AuthenticationError = errors.New(messages.AuthenticationMissing)
+	}
+	regex = regexp.MustCompile(`Token scopes: (.+)`)
+	for _, line := range lines {
+		matches := regex.FindStringSubmatch(line)
+		if matches != nil {
+			parts := strings.Split(matches[1], ", ")
+			if slices.Contains(parts, "'repo'") {
+				break
+			}
+			result.AuthorizationError = fmt.Errorf(messages.AuthorizationMissing, parts)
+		}
+	}
+	return result
 }
