--- internal/forge/gh/verify_connection.go
@@ -1,57 +1,0 @@
-package gh
-
-import (
-	"errors"
-	"fmt"
-	"regexp"
-	"slices"
-	"strings"
-
-	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
-	"github.com/git-town/git-town/v21/internal/messages"
-	. "github.com/git-town/git-town/v21/pkg/prelude"
-)
-
-func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
-	output, err := self.Backend.Query("gh", "auth", "status", "--active")
-	if err != nil {
-		return forgedomain.VerifyConnectionResult{
-			AuthenticatedUser:   None[string](),
-			AuthenticationError: err,
-			AuthorizationError:  nil,
-		}
-	}
-	return ParsePermissionsOutput(output)
-}
-
-func ParsePermissionsOutput(output string) forgedomain.VerifyConnectionResult {
-	result := forgedomain.VerifyConnectionResult{
-		AuthenticatedUser:   None[string](),
-		AuthenticationError: nil,
-		AuthorizationError:  nil,
-	}
-	lines := strings.Split(output, "\n")
-	regex := regexp.MustCompile(`Logged in to github.com account (\w+)`)
-	for _, line := range lines {
-		matches := regex.FindStringSubmatch(line)
-		if matches != nil {
-			result.AuthenticatedUser = NewOption(matches[1])
-			break
-		}
-	}
-	if result.AuthenticatedUser.IsNone() {
-		result.AuthenticationError = errors.New(messages.AuthenticationMissing)
-	}
-	regex = regexp.MustCompile(`Token scopes: (.+)`)
-	for _, line := range lines {
-		matches := regex.FindStringSubmatch(line)
-		if matches != nil {
-			parts := strings.Split(matches[1], ", ")
-			if slices.Contains(parts, "'repo'") {
-				break
-			}
-			result.AuthorizationError = fmt.Errorf(messages.AuthorizationMissing, parts)
-		}
-	}
-	return result
-}
